Workaround for NAPI weak ref implementation
The NAPI weak-ref callback does not reset the persistent handle
in the weak ref callback. This causes v8 to assert since the persistent
still has a handle to the underlying object, even after it has been
destroyed. This was previously not done because it was expected
that napi_destroy_weakref would reset the persistent before deleting it
but that isn't necessarily the case since the persistent is weak, so the
underlying object can get collected before the user gets a chance to call
napi_destroy_weakref.